notify @github/octokit-maintainers about changes in REST APIs (#15960)
We had the same thing to the last docs repository (https://github.com/github/internal-developer.github.com/pull/4800), it was very helpful. Would that be possible? I think the overhead to this solution is to add members from @docs-engineering to @octokit-maintainers, so that no review is required from me or other future Octokit maintainers

Co-authored-by: Zeke Sikelianos <zeke@sikelianos.com>