use ed25519 keys instead of rsa keys (#362)
* use ed25519 keys instead of rsa keys

RSA keys are outdated. Even by using larger 4096 keys maybe it's best to move away from rsa keys altogether. 

This PR updates the document to suggest using ed25519 keys which, at this time, defaults to a 16 round key deviation function. It may be useful to suggest more rounds for the KDF but that's for a different PR. 

--timball

* Add RSA fallback and change id_rsa to id_ed25519

Co-authored-by: hubwriter <54933897+hubwriter@users.noreply.github.com>
Co-authored-by: hubwriter <hubwriter@github.com>