Update about-github-pages.md (#734)
Lately i'm working on an unofficial Chinese translation of GitHub Pages docs due to the incompletion of the original and currently there is no way to contribute translations to the docs. So,  I hope that's ok to you guys.  

The thing is, I'm seeing that this document is using both `<user>` and `<username>` as subdomain pronouns among the context. It's not a big problem, but in my opnion, this might cause some unnecessary confusion to some users.   

If it was intended to be so, I'd like to know why?  
Thank you!

Co-authored-by: hubwriter <hubwriter@github.com>