Update workflow-syntax-for-github-actions.md
issue #825.  Under the section jobs.<job_id>.steps.run in the subsection Exit codes and error action preference: Under the bullet point for cmd, the second item starts out "cmd.exe will exit with the error level of the last program it executed, and it will and return the error code to the runner." The second "and" was deleted because it was extraneous.

The last sentence in the section jobs.<job_id>.steps.with.entrypoint read "The entrypoint keyword is meant to use with Docker container actions, but you can also use it with JavaScript actions that don't define any inputs." "to use" was changed to "to be used".