Guide to use branches instead of tags for serviceable releases
Git tags are not safe to redefine. An existing git clone *does not download tags that are already known* because git assumes that *tags will never be redefined*. 
As it stood, this doc sounded like it advocated *moving tags*, which can cause lots of confusion in a team with many git clones.
With this change, I revise the use of tag/branch terms to make it clear that branches should move and tags should remain static.