Update PATH prepending with practical example
This should save others time as I doubt anyone wanted to added the foo bar example but most wanted to add a folder which is in user home directory, and the ~ expansion does not work in this case. Also this happens to be by far the most likely PATH anyone would want to add.

In fact many would wonder why is not already in the default list.