include 'shell command' as a possible step task
This page helped me greatly to understand all the component parts in GitHub Actions. One thing that is now clear to me is that there are two types of "task" that a step can actually be - either an action, or a shell command. This is explained in a couple of places in this page, but is notably missing in this "Understanding the workflow file" section.