Clarify access level needed for secrets in web interface
Closes #1087

I considered changing the `permissions-statement-secrets-repository`
reusable to include a reference to the API, but then I noticed that the
other place using it (["Enabling debug logging"][1]) already mentioned
the API, so instead I added a note. Including a mention of "web
interface" lead to (IMO) too much duplication in text, so I rephrased it
to be more like the `permissions-statement-secrets-api` reusable.

[1]: https://docs.github.com/en/free-pro-team@latest/actions/managing-workflow-runs/enabling-debug-logging