docs: add note about dangling head
Hello!

Thanks for incorporating this change. I didn't know fo a more direct way to provide this feedback to the team, so figured this might be a good way to raise visibility. I don't expect this to get merged.

It'd be 100% better if the popup on the repo showed the user what to do here and incorporated either this or a similar command to support a seamless rename.