updating doc with links for licenses
The experiance for new users is not very obvious as to `where` they will receive their license. I've included links for Download and Install to help the user find the information.  I will be adding another commit when I have fleshed out the various ways that users get the GHES license.