Update adding-organizations-to-your-enterprise-account.md
The existing verbiage creates an absolute when noting "You cannot transfer existing organizations into an enterprise account on your end."  This is incorrect as Sales Operations has a process that they can use to assist with transfers as long as the organization is verified to be owned by the same customer. The updated tidbit provides more clarity and accuracy.