Fix .env examples for `PRIVATE_KEY` -> `GITHUB_PRIVATE_KEY` variables
This fixes naming for a couple of variables in the `.env` examples in this document.  Currently there are examples which suggest naming the private key variable `PRIVATE_KEY`, when really it should be `GITHUB_PRIVATE_KEY`.  When following along in the example and attempting to run the server, using the incorrect name in the `.env` file leads to an error.