"Issued at" time should allow for clock drift (#4156)
If `iat` is in the future from the server clock perspective GitHub will report an error. 
This can happen if the clock on the client system is significant out of sync with UTC and/or server time.

Adjusting `iat` to an arbitrary 60 seconds in the past means the error will not occur in cases where the client system clock is only off by a few seconds, which is not uncommon.