Remove the non standard `Status` header from example responses
This header was present in older Unicorn versions, but we've since
updated Unicorn and that removes the custom non standard header.

Including it in the docs implies that it might be a standard thing /
something to rely on which is not the case here. Removing it makes it
clear that we no longer send this header. Since it's a non standard
header, ideally this would have never been present in the docs in the
first place, but that ship has long sailed so all we can do is clean it
up now.