add additional information on dealing with ~/.ssh/config errors on Mac
Given the many upvotes on this post on StackOverflow https://stackoverflow.com/questions/47455300/ssh-config-bad-configuration-option-usekeychain-on-mac-os-sierra-10-12-6 it seems like many users have an issue with Mac not dealing properly with the `UseKeychain` config in the `~/.ssh/config`. This commit adds additional information on how to fix the problem and allows GitHub users to find a solution in the GitHub documentation when searching for the error message.