Clariry that App ID !== Client ID
When implementing this I made the mistake of assuming that the Client ID would be the same as the App ID. A few frustrating hours and a support ticket later I was pointed out this important difference. This commit introduces a small note about this difference. Hopefully saving others a few hours of trying to figure this out.