Korean translation of "creating-a-github-pages-site.md"
This is the Korean translation that I made for "creating-a-github-pages-site.md" section. As this is the first draft, it could be possible that I can upload the updated version if you accept. I am not sure whether or not you are accepting Korean translation too, but if you are willing to open for supporting Korean language support, then I'd glad to help you. 

Just for a note, please refer to the following work process that I had:
1/ I did not touch any {% %} part in order not to break the original document structure. 
2/ I used the formal tone of Korean language. I did not use any "friendly-tone" of Korean language. 

Please let me know if you have any comments for this. Thank you.

Best,
Donghoon Han