Detail where each context can be used within a workflow (#19900)
* Detail where each context can be used within a workflow

* tweak

* Update cross-ref to new information

This reusable is used in two articles: 'Context and expression syntax for GitHub Actions' and 'Environment variables'. In the first of these the current link point to the same section that the link is added to (which is not useful). In both articles I think the new section about context availability is a more appropriate link.

* Update content/actions/reference/context-and-expression-syntax-for-github-actions.md

Co-authored-by: hubwriter <hubwriter@github.com>

* Update content/actions/reference/context-and-expression-syntax-for-github-actions.md

Co-authored-by: hubwriter <hubwriter@github.com>

* Update content/actions/reference/context-and-expression-syntax-for-github-actions.md

Co-authored-by: hubwriter <hubwriter@github.com>

Co-authored-by: Matt Pollard <mattpollard@users.noreply.github.com>
Co-authored-by: hubwriter <hubwriter@github.com>