Clarify that `always()` has a side effect
One straight-forward interpretation of the previous wording was that the function simply always returns `true`, with no other side effects. A reasonable conclusion would then be that, in an `if` expression, `&& always()` is redundant, and can be omitted. However, this in incorrect, as the function does have the side effect of forcing the step to be executed in addition to returning `true`.