docs: add docs for dealing with special characters in ref names
There are a few places in the UI where we provide instructions for how
to do things on the command-line. We can provide shell-friendly escaped
versions of _most_ branch names, but there will always be some
edge-casey ones (eg. names containing single quotes) for which the
escaping rules vary subtly by shell. Rather than try to cram a dozen
variants of the commands into the UI, we want to direct people to the
docs for such problematic names.

The intent here is not to document every shell's rules (doomed to be
incomplete and become stale over time), but rather provide a flavour of
what users should be aware of, and some helpful links to jumping off to
official upstream docs for commonly-used shells.

Closes: https://github.com/github/coding/issues/2388