Clarify who can start a discussion on a project
I'm proposing this change because I cannot tell who can start a new discussion.  The existing 'anyone with access to the project' is ambiguous:  Does that mean anyone who can view the project?  Can people view anonymously and start a discussion?  Is it actually only people with write permission to the project who can start discussions?

My submitted text change is my guess as to how it works.  If its wrong, I'd love to correct it and add the clarification.