Reorder sections to make it more clear to the reader that requesting a CVE number is optional
At the moment the publication doc leads with the topic of requesting a CVE number.
This PR moves things around and adds language to imply that a CVE number is not required in order to publish a GHSA.
The impetus for this PR comes from an out of band conversation with the rustsec community and some confusion around GHSA publication.