Update and rename responsible-disclosure-of-security-vulnerabilities.md to coordinated-disclosure-of-security-vulnerabilities.md
Remove the word "responsible" from vulnerability disclosure policy, replace with "coordinated".

For lots of the discussion and points here I'll be referencing this article [Coordinated Vulnerability Disclosure: Bringing Balance to the Force](https://docs.microsoft.com/en-us/archive/blogs/ecostrat/coordinated-vulnerability-disclosure-bringing-balance-to-the-force) by [Katie Moussouris](https://twitter.com/k8em0) from 2010.

"Responsible" implies a set of ethics and morals are involved in vulnerability disclosure. "responsible" is also very subjective. From the perspective of the security researcher, the company (in this case GitHub) may be the "irresponsible" party for releasing the vulnerable code without adequate testing. From the perspective of the company, in this case GitHub, the researcher may be "irresponsible" if they choose to go public with a vulnerability before the vulnerability is fully patched and shipped to customer. As such, much of the industry has moved away from the term "Responsible" to the term "Coordinated".

> Responsible Disclosure should be deprecated in favor of something focused on getting the job done, which is to improve security and to protect users and systems.

Additionally, Coordinated Vulnerability Disclosure (CVD) has been the norm at Microsoft for 11 years. GitHub should similarly match the language around vulnerability disclosure that Microsoft has.