Set Password in Incognito or Private browser Mode
Please confirm this statement before approving, of course 🙇 

In addition to using the incognito mode to `reset` the password, the first admin should also use incognito to `set` the password the first time. If they do not ,and are already signed into GitHub on an existing user account, the set up process will fail and force the user to reset the passwords again.