Use in_progress workflow_job example, not queued (#22880)
* Use in_progress workflow_job example, not queued

### tl;dr

This changes the example `workflow_job` payload to be an `in_progress` event instead of `queued` event.

### why

In #22174, I added some keys to the example workflow_job json. This was correct, but by adding non-null _values_, **I introduced a subtle bug**.

It's a bug because a `queued` workflow is—by definition—_not_ yet assigned to a runner. So all the `runner_*` attributes will be `null` in the real world. 

After the workflow becomes `in_progress`, all the values are guaranteed to exist. So let's change the example to be an `in_progress` event, since it's a better example for documentation.

* Explain more about the runner_* attributes in "workflow_job_properties.md"

* Consistentize naming

Co-authored-by: Martin Lopes <martin389@github.com>
Co-authored-by: hubwriter <hubwriter@github.com>