update name of secret
When attempting to follow the guide and set a value for a secret called `GITHUB_SECRET`, the UI presented the following error:

"Failed to add secret. Secret names must not start with GITHUB_."

If this is not a valid naming convention for secrets, then we shouldn't use it in the documentation.