Change order of steps in create-translation-batch-pr.yml (#23210)
The new order will use "reset known broken files" as a last step. That
way we can better understand if the previous steps were enough to catch
those errors. The least files are reset in that last step, the better.

Also, the check for broken/mismatched liquid tags is performed after
rendering and parsing check. That helps us to identify how useful the
mismatch check is, if at all.