Update forming-calls-with-graphql.md
This commit may only make sense if you take a look at the Personal Access Token page on GitHub (where you create a new one), and open that alongside this small diff. 

----

I feel like I must be wrong / missing something, but I couldn't understand why "repo" was in this permissions list along with a subset of children under it. My best guess is that it was included on accident (or maybe some explanatory comment or text was lost, pointing out that it could be "repo" _or_ "public_repo,repo_deployment,repo:status" ...)

I'm pretty sure it's supposed to be this way (without the coarse permission of "repo"), and I'll test that shortly.