relased-Action is not triggered for drafts
My tests lead me to the assumption, that the `released`-action is only triggered when a release is done (either directly or by finishing a draft).

The docs say, `release` is also triggered when a release-draft is published. But IMO this isn't the case. And it wouldn't make sense with an action called `release`, as a draft is NOT a release.