Create github-community-guidelines.md
These are only guidelines; they do not modify our [Terms of Service](/articles/terms-of-service/) and are not intended to be a complete list. GitHub retains full discretion under the [Terms of Service](/articles/terms-of-service/#c-acceptable-use) to remove any content or terminate any accounts for that violates our Terms on Acceptable Use. These guidelines describe when we will exercise that discretion.