workflows: Add link to skipped action workaround
Actions that are skipped because of path filtering remain pending as
required on pull requests if they are set as required checks. Since the
required checks branch protection feature does not have a notion of
unnecessary checks a workaround is required.

This workaround is documented but hard to find so I added a note
pointing to it from the section that introduces path filtering for
workflows.

Fixes #16048