Remove redundant mention of "Allow edits from maintainers" setting from intro
The very first sentence of the introductory paragraph of the "Creating a pull request from a fork" page is about a "Allow edits from maintainers" setting that will only be available after step 5 of the instructions.

This setting is already well explained at the appropriate place in the instructions. Mentioning it redundantly at the very start of the document is confusing and makes it less approachable to the target reader.