Add note about how to specify event
Hi, I'm new to GitHub action and I would like to push the built docker image to the docker hub when I commit. I pasted the demo yml to try it but I was blocked and spent a lot of time solving it, because our doc doesn't note that we need to specify the event that triggers the docker action. 
So I add a note in our docs and hope it could help the other newer.