Add `repo.download_zip` action to audit log docs
This updates the audit log documentation to include the `download_zip`
action under the `repo` category which is currently missing.

This is triggered when a user clicks the "Download ZIP" link in the
repo UI or uses the "Download a repository archive" APIs (which
support both ZIP and TAR archives).

Someone has [asked][1] about this on the Community Forum recently. I
suspect this is related to the recent Heroku incident which drove people
towards the audit log. As has been [noted][2] in Slack, the Heroku
GitHub integration does like to frequently download ZIPs.

[1]: https://github.community/t/github-audit-log-doesnt-mention-repo-download-zip/232978
[2]: https://github.slack.com/archives/C03BSQZ89NG/p1650146666951349?thread_ts=1650129385.654679&cid=C03BSQZ89NG