docs: do not suggest updating old tags
The security guide recommends users avoid using tags when using third-party actions.
See "Pin actions to a full length commit SHA" and "Pin actions to a tag only if you trust
the creator" on the following page:
https://docs.github.com/en/actions/security-guides/security-hardening-for-github-actions#using-third-party-actions

But this doc suggests action authors provide semantic versioned tags, that will
motivate action users to use tags instead of full-length commit SHA.

To keep documents consistent, here we suggest semantic versioned tags only for internal-use actions.
It is OK for internal-use actions, like private actions used for audits because the author is not
a "third-party" for users and it should be trusted for them.

Signed-off-by: Kengo TODA <skypencil+github@gmail.com>