fix #18528 : Remove `-` from `no-reply`
Remove the erroneous `-` from instances of `no-reply` in 'setting your
commit email address' doc, including reusable. This results in `noreply`
which is the format the noreply email address is presented in.

A search through all files returns other instances of 'no-reply', some
in code style, some not. I will have a read through the relevant docs
and see if these should be changed but am uncertain about some at first
glance as I/my machine seem to meet `if` criteria that negates the
rendering of this wording. Adding them as a separate commit seems more
logical as it allows dropping those changes as they may be deemed not
related to #18528. Hope that is an acceptable practice!