Recommend an `Accept` header without V3 in it when documenting our REST API operations (#28477)
* Recommend an `Accept` header for REST API requests which doesn't mention V3

In the documentation for every REST API, in the "Headers"
section, we say for the `Accept` header that:

> Setting to `application/vnd.github.v3+json` is recommended.

Putting the V3 bit in doesn't actually do anything - in fact,
a high proportion of API consumers don't do it - and this
recommendation is going to be confusing once we launch our
new API versioning. This recommends that people use
`application/vnd.gitrhub+json` instead.

We could just recommend `application/json` - I'd be in favour
of that! - but I'm trying to make this change as small as
possible.

* Update cURL and GitHub CLI code examples to use an `Accept` header without V3

Co-authored-by: Sarah Edwards <skedwards88@github.com>