Add context about how contribute page is created
Didn't know about the contribute page and was wondering how to create it. Took me a bunch of searching to finally find the info at https://docs.github.com/en/issues/using-labels-and-milestones-to-track-work/managing-labels. 

I think this context helps keep people on this page and focus on the point that the contributing guidelines show up on the contribute page.