Update best practices to positive statements
Most readers don't expect "don'ts", "nots" or other negatives in instructions. When we're giving a list of best practices, we should advise people what *to* do, rather than what *not* to do. The last bullet point was a little tricky for my brain to write fully in the affirmative, thus the longer sentence.