Documentation on how to use the release trigger is wrong and confusing.
### Context

The documentation does not correctly describe how to make a release that triggers when you publish a release, or change the toggle to prerelease and update the release. To do this you need to use both the `released` and `prereleased` option. This contradicts what was written in the deleted note.

### Changes

- Deleted note that says you should use `published` to trigger when a prerelease or a release is happening. This is wrong. With `published` it only triggers the first time a release is published.
- Added info on how to trigger workflows with release in typical scenarios, e.g., when you you want to trigger each time a prerelease or release is published, or edited from prereleas to release or vice-versa.