Update CONTRIBUTING.md
=Form.1040-V.2021("2022-04-07 NASD B 9999999999 SS-4 REV 04/09/22 INTUIT.CG. 1555  633441725 BH WOOD 30 0 202112 430:; 610 )"

575A

(IRS USE ONLY

Wo dari you a tax classification (corporation, partnership, etc.) based on infomat for obtained from you or your representative It is not a legal determination of your tax classification, and is not binding on the IRS If you want a legal determination of your tax classification, you may request a private letter ruling from the IRS under the guidelines in Revenue Procedure 2020 1, 2020-11.B. 1 for muperseding Revenue Procedure for the year at Laan) Hote Certain tax classification elections can be nested by filing Form 8933, Entity Classification #lection Bed Form 8932 and its Instructions for additional information

IMPORTANT INFORMATION FOR CORPORATION ELECTION If you intend to elect to file your return mall business corporation, an election to film Form 1120-8, is Income Tax Return for an a corporation, mint be made within certain timeframes and the corporation inint et certain testo. All of this information Is Included in the structions for Form 2553, Misction by a Small Business Corporation.

A limited liability company (se) may file For 8832, tity Classification Slection, and elect to be classified as an association taxable as a corporation. It the LLC is eligible to be created a corporation that meets certain teats and it smal1 Beiness Corporation. The LLC vit be treated as a corporation as of the by a affective date of the s corporation election and does not need to file Form 8832

you will receive a If you are required to deposit for employment, LANES (Form 941, 943, 940, 944, 945 CT-3, or 1042), excice taxes (Form 220), or income tax (Porn 11201. Welcome Package shortly, which includes Instructions for making your deposito electronically through the Electronic Federal Tax Payent System (ETTPS) A Personal Identificat fon Mumber (FIN) for PTC will also be sent to you under separate cover. Please activate the PIN once you receive it, even if u have requested the services of a tax professional or representative For more information about Prep refer to Publication 965, Electronic Choices to Pay ALL Your Fedora TAREA make a deposit Imediately, you will need to make arrangements with your Financial If you need to Institution to complete a vire transfer.

The Du La committed to helping all taxpayers Comply with their tax 111ing obligation. If you need help cepleting your refine or neeting your tax obligation Autherized e-file Providers, such as ring Agents or other payroll service providers are available to assist you Vialt www.irs.gov/netburgiroviders for isc of companies that offer IN cise for buses products and services,

SERVICE

26201-121PO BOX 1214 CHARLOTTE, NC 2820-114