Clarify documentation for `versioning-strategy` option
We receive the occasional support request and there are also multiple issues filed against https://github.com/dependabot/dependabot-core/ asking about "I tried X strategy for Y ecosystem but it's not working".

So rewrite this sections of the docs to add some clarity:
1. Add table illustrating how the different strategies work for a given scenario (🎩 💁‍♂️  to @jonasfj from https://github.com/dependabot/dependabot-core/issues/4979#issue-1196967453)
2. Invert the table from "strategy -> supported ecosystems" to "ecosystem -> supported strategies" because our users arrive with questions about a given ecosystem, not a given strategy
3. Add additional ecosystems to the list
4. Shorten the example because we only need one, additional ones just contribute to the #walloftext feeling
5. Re-order everything lexicographically

Hopefully this will provide better documentation that we can point users to in response to questions.

Co-authored-by: Jonas Finnemann Jensen <jonasfj@google.com>
Co-authored-by: mc <42146119+mchammer01@users.noreply.github.com>