fix: do not include `$` in command
Currently, the command includes the `$` terminal prompt. When you click the copy to clipboard button, this is also copied to the clipboard. The command will fail if you paste and hit return in the terminal. I have also seen this on a few other pages. I will have a look and open an issue to highlight those.