Cleanup docker registry docs
Dependabot now supports:
1. Public AWS ECR: https://github.com/dependabot/dependabot-core/issues/4212
2. Microsoft Azure Container Registry: https://github.com/dependabot/dependabot-core/issues/3689
3. Any registry that uses HTTP Basic Auth instead of the OCI-spec-compliant central token service: https://github.com/dependabot/dependabot-core/issues/3689#issuecomment-1272037775

So now that we support all the major docker/container registry providers, there's no need to individually list them... this way we don't have folks wondering "my registry wasn't explicitly listed, is it supported?"

So switch to mention the two authentication schemes that we support and leave it at that.

Note: I left the `ecr-docker` example, as that is a one-off workaround we added to our internal code a long time ago because ECR isn't OCI compliant. So still needs the custom key to indicate the workaround should be used. Long term I hope ECR migrates to a more standard auth flow and we can eliminate that custom workaround, but they're not there yet.