Update searching-for-repositories.md
The search **jquery in:name** will only find repositories that **start** with jquery. 

https://github.com/search?q=jquery+in%3Aname   (122,150 repository results)

For example if one change the search to **query in:name**, i.e. drop the j, then is if the original description in this file was correct it should find more repos. However as one can see by running this query

https://github.com/search?q=query+in%3Aname 

you get less results (81,119). This is because the query is not actually searching if the word "query" is in the name but rather does the repo start with the word "query".