rm misleading doc about paths and paths-ignore
I would love for this PR to include a suggestion of better content here, but I don't know that it exists. Also, I would be open to editing to just add a large caveat about the problems with this approach.

In short, this approach is un-safe because `paths` and `paths-ignore` don't work the way these docs suggest then do. From the [docs for paths-ignore](https://docs.github.com/en/actions/using-workflows/workflow-syntax-for-github-actions#example-excluding-paths):

> When all the path names match patterns in paths-ignore, the workflow will not run. *If any path names do not match patterns in paths-ignore, even if some path names match the patterns, the workflow will run.*

(emphasis mine)

This means that if you change any file not listed in `paths` or `paths-ignore`, *both workflows will run*. 

Running both is probably fine, except for a few things:

1. it violates user expectations - at least, that was not my expectation from reading this section of docs
2. it will hide real failures. For example if both workflows run, the "skip" one succeeds, but the "real" one fails due a parsing error, you will get a successful status check on your PR (tested - this is how I went down this rabbit hole)
3. it is a race condition - what if you have auto-merge turned on? The "skip" workflow is almost certain to finish faster leading to a merged PR before the other workflow is done. (not tested)

The only alternative I have heard (from a coworker) is to use `paths` and negation (`!`) instead of `paths-ignore`. I don't think that would work because `paths-ignore` is an OR not an AND. You need to negate the whole list, not each item individually (!(a || b) rather than (!a || !b)).