Add `strategy` to keywords for jobs that call workflows
Buddy and I brainstormed using `matrix` values when it came to calling our reusable workflow and we were bummed to discover it wasn't supported. Low and behold [it is](https://github.com/qubitz/workflow-testbed/actions/runs/3388522235).

Besides just adding `strategy` to the list of supported keywords, I figured I'd throw it into the example as well. The use in the example is a little contrived, suggestions welcome. Or maybe leave it out of any example?