Add useful info to "pushing a container"
When you push a container to ghcr.io without having the label `org.opencontainers.image.source` set GitHub won't connect the image to a repository. This results in the `GITHUB_TOKEN` not having the appropriate permissions.

Example:

Repository url: `flopana/awesomeRepo`
I push an image that is tagged with `ghcr.io/flopana/awesomeRepo:latest` to ghcr.io, GitHub won't link these two together.

If I now use the `secrets.GITHUB_TOKEN` in GitHub Actions on the repository `flopana/awesomeRepo` and try to push a container with it it fails with `403 Forbidden`. Because the token belonging to this repository doesn't have the permission to write to a package that is not belonging to this repository.

If there is no package under that namespace and you push an image from the pipeline with the github token to ghcr.io, Github will link these two even without the label set.

But by advising the user to set the label in the Dockerfile github will link both together regardless of the order of pushing images.

My changes in the Documentation should prevent users from running into this problem like I did.

Feel free to make changes to my grammar/wording since English isn't my mother tongue.