Use a random delimiter in examples
It was a bit jarring seeing a strong warning about choosing random delimiters, then an example immediately afterwards that didn't follow that advice.  This fixes that.

I'm not 100% confident with `base64` here.  The inclusion of "/" and "+" characters should be fine, but sometimes they are treated specially in ways that might not produce good results.

Also, my powershell isn't that strong, but this seems to do the trick (with slightly lower entropy as I don't include "+" and "/").

FWIW, 120 bits of randomness (or close to that) should be plenty.  Using a multiple of three bytes ensures that `base64` doesn't add any awkward "=" characters.  I'd use hex, but then I'm less confident in the availability of `xxd`.