Updating the titles for clarity.
Part of the reason these docs needed to be updated was to make these titles/steps more specific. Oauth App Integration in OKTA ≠ GitHub Oauth App, and specifying that this is being set up for SAML -- specifically for a GHEC organization (and not for an enterprise) -- is very important in providing clarity.