Mention removal of email from a secondary account.
In my case my email was associated to another personal account, not owned by me. It probably was because a side effect of the transfer of ownership of a repository, I believe from a collaboration to a personal account. After contacting the owner the e-mail was removed and the issue was immediately resolved (even past merge requests were thereafter attributed to the right personal account.)