Update writing-mathematical-expressions.md
This issue is about displaying math blocks inside of the `details` blocks (and potentially other HTML blocks) regardless of the type of math block. This can include blocks such as 

```
$$
\begin{pmatrix}
  0 & 1 \\
  2 & 3
\end{pmatrix}
$$
```

or blocks written like this: 

~~~
```math
\begin{pmatrix}
  0 & 1 \\
  2 & 3
\end{pmatrix}
```
~~~

This issue is currently blocking the use of math blocks inside of `details` blocks, as well as other HTML blocks. This limitation prevents users from taking full advantage of these blocks, and should be noted in documentation. To ensure users are not held back, we must actively search for a potential solution to this issue so that math blocks can be displayed in these blocks.