Update understanding-the-search-syntax.md
The current query does not appear to work: https://github.com/search?q=cats+pushed%3A%3C2012-07-05&type=Code&utf8=%E2%9C%93

It searches for the string 'cats pushed:<2012-07-05', rather than treating pushed:<2012-07-05 as a qualifier. 

Perhaps this is because this pushed/created date/time qualifiers are no longer supported for Code type searches? If so, I suggest we update this example to be a Repository type search instead, as reflected in this PR.

However, if date/time qualifiers are supported for Code type searches then I'd be happy to update this pull request with a working example.