Navigation in doc about concepts faster
To make navigation between concepts in documentation faster. For new users who read about build and publish but need a refresh on the concept of starter workflows. A direct link to relevant documentation is proposed to be added. Similar acitvity could be done under the "Build&Test" examples for other languages.